+++ lib/types/Storage.ts
@@ -61,7 +61,6 @@ export interface StorageOptions extends CookieOptions {
   storageTypes?: StorageType[];
   storageProvider?: SimpleStorage;
   storageKey?: string;
-  useMultipleCookies?: boolean;
 }
 
 export type StorageType = 'memory' | 'sessionStorage' | 'localStorage' | 'cookie' | 'custom' | 'auto';