+++ test/spec/TokenManager/browser.ts
@@ -105,7 +105,6 @@ describe('TokenManager (browser)', function() {
         storageManager: {
           token: {
             storageTypes: ['cookie'],
-            useMultipleCookies: true,
           }
         }
       });